Parallel safe version of VE notebook using mpi4py
This is probably not the best solution, but the analysis previously
assumed that the sample tracer was in the current section of the domain
and this generally will fail. My solution is not that great if there
are several particles in different domains … it’s just a fix for this
one case.