I tried to add some docstrings for the analytic functions ...
BUT I can’t really make any connection to the analytic solutions that
are in the underworld repo / documentation. And there is no
documentation at all of the ones in the hpp file.

All this because they are not properly described in the paper that is
supposed to cover all of this.

Can somebody dig a pit, cover it with some sticks and grass and pop a
slab of tooheys red on the far side ?