An attempt to streamline the solver code
And to go through the code to work out what is not needed and what is
duplication. There is much to do in this but best to go slowly and not
break anything.

I have cached velocity which solves the RHS term in the SCR solver.
This allows the previousGuess to work effectively. At the cost of one
extra vector of storage this is a good trade-off. Not sure how to free
this vector at close-down time.

The DGMGD penalty is not much use and I think the implementation should
be dropped as it has no obvious benefit to the solver speed or
robustness. Suggest we replace this by allowing \lambda to be a
user-defined function if  the user wants the penalty to vary.

There is a clash between bfbt and penalty which should perhaps be
detected and disallowed.

All tests are passing !