Adding a small modification to activate the non-linearity in the 2nd viscosity
The non-linear iteration needs to be told how to find the non-linear
terms otherwise it won’t activate the non-linear iterator. Here I have
activated it on the 2nd viscosity for velocity and pressure. I have not
exposed these terms in the interface, I tunnel down through the SLE.
Preliminary testing shows this working, but will add a proper test soon