Added functionality to define a vector function using a list of subfunctions. for example: 	vecfun = (fn_x, fn_y, fn_z)
If a user defines this in python, it will only exist as a python tuple of
functions, though it will be converted under the hood.  however, it seemed
it may be useful for users to directly convert this to a native function,
probably only so that they can access the 'evaluate' method.  hence, i've
renamed the private method '_CheckIsFnOrConvertOrThrow' to simply 'convert',
and added docstrings/doctests etc.