Adding nonLinearKillNonConvergent and nonLinearMaxIterations to the stokes solver interface
I think the nonLinear kill should be False by default since it calls
abort() and kills the kernel in a notebook. This is not good behaviour !

solver.solve( nonLinearIterate=True, nonLinearTolerance=0.001,
nonLinearKillNonConvergent=False,nonLinearMaxIterations=5,
print_stats=True )

We should make a convergence check somewhat simpler to do for the user.