BUGFIX fixes #204
I have moved some of the function testing (for conditional functions) to be
performed during the heavy stage of evaluation. This should not incur any
appreciable performance overhead.

The reason for this is that with the introduction of non-global swarms,
some functions now may not be defined across the entire domain. The initial test
phase however only tests at a single (somewhat) arbitrary location, and hence is
not appropriate for this usage.  Instead we ensure correct configuration during the
heavy part of evaluation, where we are likely to encounter all functions which
are part of the conditional function.