The addition of @property attributes to the stokes object can't be added conditionally
This was done for fn_one_over_lambda and fn_source so that they would
not be modified / added in the case that the underlying PETSc object
was not defined but the implementation did not work at all. A simple
fix is to add the appropriate attributes but lock them if the
appropriate matrices have not been defined. It’s not a pretty but seems
to work. More elegant solutions welcomed