Updates to the NeumannCondition
The usage and syntax of the NeumannCondition object has changed.
'flux' --> 'fn_flux'
In the systems: SteadyStateHeat & AdvectionDiffusion, fn_flux now takes a scalar value for heat flow.
In the the stokes system fn_flux now takes a traction vector.

The calculations for the heat flow or traction vector is up to the user. More on this to come.

For now see the updated examples
docs/test/SteadyState.ipynb, docs/test/uplift.ipynb, docs/test/StokesEq_PoiseuilleFlow.ipynb
for a guide.