Updates to speed up swarm checkpointing
Implementation of mansour's suggestion to add the proc particle counts
to the swarm checkpoint - and then use that to optimise the swarm reload.

Upon testing, up to 8 cores with 67e6 particles. roughly 50% speedup.
More testing to come.
Note if different no. procs used the algorithm doesn't use the optimisation option.