+ 'try_optimised' parameter to swarm load
The 'try_optimised' (true by default) enables the swarm to be loaded with
processor offsets. This can speed up the load, but can also misread
the swarm if a different resolution and/or different proc partitioning is used.

So 'try_optimised' can be set to false, causing the original brute force
method to run.

Bounding extremes check should be implemented on this next.