Big updates for SafeMaths and also wider error messaging.
We now record the stack at function construction time which is then reported
back to the user when an error relating to the function occurs. This
addresses the disconnect that occurred due to functions usually being
constructed, and then tested much later. The testing still occurs later,
but now when an error occurs, it points the user to exactly which function
definition caused the error.

This is currently only used in SafeMaths, though it will also be rolled
out to other functions which can report errors. Compounded SafeMaths are
also now should operated correctly, and point users to the correct SafeMaths
definition raising the floating point exception.

Also some house keeping, with SafeMaths & CustomException implementations
moved into their own files.