These changes address the disconnect that exists between Function construction and Function initialisation (and also usage).
Functions are constructed agnostic of their eventual input, and therefore it is not possible to test (at construction time) if the eventual input is compatible. As such, a user constructing a Function in a location very far (script position wise) from the eventual initialisation (which occurs at `solve()` time for example), will often find it difficult to trace errors back to the offending Function's definition (ie construction). To address this, we now record the Function's construction time Python stack information and class type. This is then displayed to the user upon any encountered (Function related) errors (either at initilisation or during usage), showing them the offending line in their script and the object's class type. Hopefully this will allow them to resolve their issues with less gnashing of teeth and rending of garments.