Slade (#339)
* getting slade solver back

* SLCN with sourceTerm and NeumannBCs + api change

SLCN can now we swapped in and out to replace the old AdvectionDiffusion
system. Source Terms and Neumann conditions work for it.
Further testing to come. More worked required on it handling non
linear situations and correctly accounting for transient source terms.

Also API change for SteadyState and existing AdvectionDiffusion systems.
I removed the -ve condition on the underlying systeam Neumann boundary
definition, a legacy factor from a while back.
I think the current implementation is a lot more straightforward.

* Adding the Cosine Hill advection test

* GLORY - a bug fix for deformed meshes

A per element width is required for the cubic interpolant required for SLADE.
Dan reckons sweet as!
The previous implmenetation used a single element width for the entire proc, this was broken.

* Extra info in error statement

* Small fixes from the merge