Initial work on implementing the FSSA stabilisation term.
The FSSA algorithm introduced by

Kaus, B.J.P., et al., A stabilization algorithm for geodynamic numerical simulations with a free surface. Phys.
Earth Planet. In. (2010), doi:10.1016/j.pepi.2010.04.007

This is only the C++ backend implementation as it stands the matrix term is yet to incorporate the disagonalisation and still
requires testing.