V2.16.x -> main (#715)
* V2.15.1 PR (#674)

* starting for new release uw2.15

* Update surfaceProcesses.py

Update timeField to record time of deposition of sediment in SP functions

* Update surfaceProcesses.py

Fix typo in 3D surface.

Also reset time of the air material to the model time

* Update surfaceProcesses.py

fixed typo in func

* updating the setonix jobs runners

* Update surfaceProcesses.py

Simplify 2D velocity and diffusivity SP functions

* Update surfaceProcesses.py

Didn't create object to bcast surface function to

* Update surfaceProcesses

Revert back to MPI gather of data over evaluate_global

Update the examples

* Adding melt dynamic heating to the advection diffusion solver
see issue #669

* docker utils updates

* Adding some final fixes to the dockerfile scripts

* Update CHANGES.md

* Remove obsolete glucifer bits

* Update dockerfile Badlands install and CHANGES.md

* Update CI for docker builds

* Fixing the velocitySurface 2D and 3D implementation to handle restarts.

Changes also include:
* Renaming the surface tracers to prevent unexpected overwrites
* The surfaceProcess class saving the passive tracer key to use during
initialisation and solve. This is to make the code not save an instance
of the pt to the swarm incase the pt is swapped in/out of Model memory.
And helps with code readability.

* Get the latest release of badlands from pypi

---------

Co-authored-by: Ben Knight <55677727+bknight1@users.noreply.github.com>

* starting 2.16.x

* Moving to a specific badlands version & (arm compatible version of) jupyter-server extension command

* update freesurf to 3D in uwg (#683)

1. update the free surface function in UWGeodynamics to a 3D version (see details in _freesurface.py).
2. add an example 1_23_05_FreeSurface_3D_Relaxation in docs/UWGeodynamics/examples/.

* First cut at the GA4 trojan.

Unfortunately GA4 isn't the same as the previous UA system by google. So location tracking is not enabled.

Every import (where the python interpreter is up for ~2 seconds) will produce a GA4 event that records:
- the version of Underworld used
- the platform it's being used on
- the number of mpi processors in the execution of python.

This event is anonymous and no user data other than the above, and a random key (assigned during install) is sent to the GA4 server.
This will provide valuable metrics to the Underworld development team.

As previous users can disable this trojan by setting the environment
variable UW_NO_USAGE_METRICS to any value, eg.
export UW_NO_USAGE_METRICS=1

Also updating the version number to 2.16.0

* Update install instructions to include ninja and python 3.11

* Update UWGeoTutorials.rst

fixing some missing links, pointed out to me by a user

* move underworld directory into 'src'. Reduce the namespace issues.

* Updating pyproject.toml and version number to support python 3.12

* Fix this test.

'imp' module is not longer in python 3.12. Moving to 'runpy' instead.

* Turn off pedantic compiler warnings. StGermain is full of pointer / int conversion code that trips up sensitive compilers

* moving everything to src/ to avoid namespace confusion

* error in prefix of install dir

* Remove this because it's causing lots of compiler warnings.

* Adding a script clean.sh to quickly strip the repo of python build/cache
files

* Fixes to use 'src' strategy to install Underworld with pip install .

* Updating cleaning script

* manually send country/region ip info with this add to the uw trojan

* Typo with white space

* Missed this update

* Geo information in G4A stats

Not the best implementation but G4A doesn't support geographic data via
it's Measurement Protocol. So the workaround ("hack") I came up with is
to put the country name in the event title.

The existing issue of G4A's lack of geo info.
https://issuetracker.google.com/issues/237566549?pli=1

* Cleaning up some warnings from compilation.

* Build changes for 2.16

1. setup.py -> pyproject.toml new standard, PEP 621.
2. all source code now under `src`.

This is a start for the new standard. Main issue is build isolation
forces CMake caching to break, so I have disabled the caching for now.

* Update CHANGES.md

* Rename ressources -> resources in UWGeodynamics

The path ressources is still symlinked for backwards compatibility.

* Misc changes for 2.16

* pyproject.toml with dyanmic versioning
* UWGeodynamics/tutorial: ressources -> resources

* simplify dependencies in pyproject.toml

* remove erroneous '-python' option - only affecting gadi

* reformating this with 'clang-format --styleLLVM'

* Big update of Dockersfiles

- More automation of versioning in `docker-build.sh`
- Moving to Ubuntu 24.04 as base OS
- Update to petsc 3.22.2
- using `podman` instead of `docker`

* force lavavu-osmesa install with compilation

* Switch to lavavu instead of lavavu-osmesa

* Just use lavavu with --no-binary :lavavu:

* Force numpy<2 in all dockers

* add an extra check if lavavu is valid - for arm64 linux installation

* Update the dockerfiles

* Updates to conda files, ie latest petsc, numpy<2, cython>=3

* Fix typo when calling bash

use 'set -x' instead

* fix typo in dockerfile

* Add pawsey container build recipe

* Updates for v2.16.x

* typo in check for lavavu

* Add environment file

* Docker updates (#711)

* Update dockerfiles

- Rebase to a python image based on Debian bookworm
- Update runtime requirements (mainly for lavavu)
- Modify bash script to allow switching between mpi implementation & change of base image

* Remove pawsey files

Can be built with the previous commit (8048d4c891d36b9bdd858a6e1fc14f23a3b4713c)

* Update CHANGES.md

* add some notes on the complex build system of UW2

* Updating the environment to include badlands. Coupling is a go.

* Create conda-build-test.yml

Use the conda/environment.yaml to provide a quick environment to compile and run the unit tests.

* Update conda-build-test.yml

Need to init the conda env to run things in the mamba activate env.

* Update conda-build-test.yml

* Update conda-build-test.yml

* Update conda-build-test.yml

* Update conda-build-test.yml

get the tests running

* Update conda-build-test.yml

* Update conda-build-test.yml

just get it working. Would like to split the steps of compiling and running the tests, but can't work out the syntax.

* Update conda-build-test.yml

Include mac-intel in tests

* fixes for docker naming and notes of building in podman

* add some documentation in the docker-builder.sh
:

* Fix spacing

* Update Dockerfile

To get mybinder links working

* Update Dockerfile

Fix warning of legacy variable UW_MACHINE

* Update README.md

Adding a new mybinder link

* Update README.md

* Dev into 2.16.x (#708)

* Fix cohesion bug by adding epsilon1 and epsilon2 (#707)

* Update conda-build-test.yml

mac-os from build runner. Its not available and that's causing errors.

---------

Co-authored-by: tiannh7 <156781848+tiannh7@users.noreply.github.com>

* Sphinx-updates (#713)

* Fix cohesion bug by adding epsilon1 and epsilon2 (#707)

* Update conda-build-test.yml

mac-os from build runner. Its not available and that's causing errors.

* Update .readthedocs.yml

* update conf.py path from "src/"

---------

Co-authored-by: tiannh7 <156781848+tiannh7@users.noreply.github.com>

---------

Co-authored-by: Ben Knight <55677727+bknight1@users.noreply.github.com>
Co-authored-by: Neng Lu <57476594+NengLu@users.noreply.github.com>
Co-authored-by: Arijit Laik <laikarijit@gmail.com>
Co-authored-by: tiannh7 <156781848+tiannh7@users.noreply.github.com>